ref: Move VideoConstraintsCompatibility to the cc package.
